requirements.txt
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask==0.10.1
MySQL-python==1.2.5
PyYAML==3.11
docopt==0.6.2
requests==2.5.0
tornado==4.0.2
